SSCS-11348 Refactor and add "refused" option
